ClusterSet controller and Common Area interface
ClusterSet controller - implement the reconciller for the ClusterSet
CRD. The ClusterSet CRD depends on the ClusterClaim CRD, although
the two can be configured in any order, the ClusterSet CRD will
perform validations to make sure it will be processed only after
the appropriate ClusterClaim CRDs are present.

Leader ClusterSet controller - Currently creates access to the
local namespace using the Common Area abstraction

 Leader ClusterSet Controller
            |
 InboundLeaderClusterManager
            |
       Common Area
  (access to local cluster
   namespace)

Member ClusterSet controller - read the leader list from spec
and mantain a connection to each of them using the controller-runtime
library. This connection is abstracted as the common area of the leader
cluster namespace for the member to write to and listen for
monitor updates

 Member ClusterSet Controller
            |
   Remote Cluster Manager
            | 1:N
      Remote Cluster
            |
      Common Area
  (access to remote cluster
   namespace)